+++ src/Resources/contao/dca/tl_pwa_pushnotifications.php
@@ -22,20 +22,18 @@ $GLOBALS['TL_DCA'][$table] = [
 		],
 	],
 	'list'        => [
-		'label'             => [
-			'fields'         => ['title', 'sendDate', 'sent'],
-			'format'         => '%s',
-			'label_callback' => ['huh.pwa.datacontainer.pwapushnotification', 'onLabelCallback'],
-		],
 		'sorting'           => [
 			'mode'        => 4,
-			'fields'      => ['sendDate DESC'],
+			'fields'      => ['sent ASC, dateSent DESC'],
 			'panelLayout' => 'filter;sort,search,limit',
-			'flat'        => 6,
-
 			'headerFields'          => ['title'],
-			'child_record_callback' => ['huh.pwa.datacontainer.pwapushnotification', 'onLabelCallback'],
+			'child_record_callback' => ['huh.pwa.datacontainer.pwapushnotification', 'onChildRecordCallback'],
 		],
+        'label'             => [
+            'fields'         => ['title', 'sent'],
+            'format'         => '%s %s',
+            'group_callback' => ['huh.pwa.datacontainer.pwapushnotification', 'onGroupCallback'],
+        ],
 		'global_operations' => [
 			'all' => [
 				'label'      => &$GLOBALS['TL_LANG']['MSC']['all'],
@@ -71,7 +69,7 @@ $GLOBALS['TL_DCA'][$table] = [
 	],
 	'palettes'    => [
 		'__selector__' => ['clickEvent'],
-		'default'      => '{message_legend},title,body,icon,iconSize;{behavior_legend},clickEvent;{send_legend},sendDate;',
+		'default'      => '{message_legend},title,body,icon,iconSize;{behavior_legend},clickEvent;{publish_legend},published,start;',
 	],
 	'subpalettes' => [
 		'clickEvent_' . \HeimrichHannot\ContaoPwaBundle\DataContainer\PwaPushNotificationContainer::CLICKEVENT_OPEN_PAGE => 'clickJumpTo',
@@ -131,22 +129,6 @@ $GLOBALS['TL_DCA'][$table] = [
 			},
 			'sql'              => "varchar(64) NOT NULL default ''"
 		],
-		'sendDate'      => [
-			'label'     => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['sendDate'],
-			'inputType' => 'text',
-			'default'   => time(),
-			'eval'      => ['rgxp' => 'datim', 'doNotCopy' => true, 'tl_class' => 'w50', 'datepicker' => true],
-			'sql'       => "int(10) unsigned NOT NULL default '0'",
-			'flag'      => 8,
-		],
-		'sent'          => [
-			'label'     => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['sent'],
-			'inputType' => 'checkbox',
-			'filter'    => true,
-			'eval'      => [],
-			'sql'       => "char(1) NOT NULL default ''",
-
-		],
 		'receiverCount' => [
 			'label'     => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['receiverCount'],
 			'inputType' => 'text',
@@ -163,14 +145,44 @@ $GLOBALS['TL_DCA'][$table] = [
 			'eval'      => ['tl_class' => 'w50', 'includeBlankOption' => true, 'submitOnChange' => true],
 			'sql'       => "varchar(255) NOT NULL default ''",
 		],
-		'clickJumpTo'   => [
-			'label'      => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['clickJumpTo'],
-			'exclude'    => true,
-			'inputType'  => 'pageTree',
-			'foreignKey' => 'tl_page.title',
-			'eval'       => ['fieldType' => 'radio'],
-			'sql'        => "int(10) unsigned NOT NULL default '0'",
-			'relation'   => ['type' => 'hasOne', 'load' => 'eager']
-		]
-	],
+        'clickJumpTo' => [
+            'label'      => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['clickJumpTo'],
+            'exclude'    => true,
+            'inputType'  => 'pageTree',
+            'foreignKey' => 'tl_page.title',
+            'eval'       => ['fieldType' => 'radio'],
+            'sql'        => "int(10) unsigned NOT NULL default '0'",
+            'relation'   => ['type' => 'hasOne', 'load' => 'eager']
+        ],
+        'sent'          => [
+            'label'     => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['sent'],
+            'inputType' => 'checkbox',
+            'filter'    => true,
+            'eval'      => [],
+            'sql'       => "char(1) NOT NULL default ''",
+        ],
+        'published'   => [
+            'label'     => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['published'],
+            'exclude'   => true,
+            'filter'    => true,
+            'flag'      => 1,
+            'inputType' => 'checkbox',
+            'eval'      => ['doNotCopy' => true],
+            'sql'       => "char(1) NOT NULL default ''"
+        ],
+        'start'       => [
+            'label'     => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['start'],
+            'exclude'   => true,
+            'inputType' => 'text',
+            'eval'      => ['rgxp' => 'datim', 'datepicker' => true, 'tl_class' => 'w50 wizard'],
+            'sql'       => "int(10) unsigned NOT NULL default '0'"
+        ],
+        'dateSent'      => [
+            'label'     => &$GLOBALS['TL_LANG']['tl_pwa_pushnotifications']['dateSent'],
+            'inputType' => 'text',
+            'eval'      => ['rgxp' => 'datim', 'doNotCopy' => true, 'tl_class' => 'w50', 'datepicker' => true],
+            'sql'       => "int(10) unsigned NOT NULL default '0'",
+            'flag'      => 8,
+        ],
+    ],
 ];